SRB2: wow, MSVC really do not support the d suffix
git-svn-id: svn://code.srb2.org/SRB2/trunk@5265 d18046d3-e9af-4449-a702-ac2b6c676ba2

+++ src/hardware/hw_main.c
@@ -4097,7 +4097,7 @@ void HWR_RenderPlayerView(INT32 viewnumber, player_t *player)
 	atransform.fovxangle = fpov; // Tails
 	atransform.fovyangle = fpov; // Tails
 	atransform.splitscreen = splitscreen;
-	gr_fovlud = (float)(1.0l/tan(fpov*M_PIl/360l));
+	gr_fovlud = (float)(1.0l/tan((double)(fpov*M_PIl/360l)));
 
 	//------------------------------------------------------------------------
 	HWR_ClearView();
